openalex-neo4j-import/requirements.txt
neo4j>=5.14.0
python-dotenv>=1.0.0
